abs/fabs -> fabsf

+++ src/in_null.c
@@ -216,7 +216,7 @@ void IN_JoyMove (usercmd_t *cmd)
 			if ((joy_advanced.value == 0.0) && (in_mlook.state & 1))
 			{
 				// user wants forward control to become look control
-				if (fabs(fAxisValue) > joy_pitchthreshold.value)
+				if (fabsf(fAxisValue) > joy_pitchthreshold.value)
 				{
 					// if mouse invert is on, invert the joystick pitch value
 					// only absolute control support here (joy_advanced is false)
@@ -243,7 +243,7 @@ void IN_JoyMove (usercmd_t *cmd)
 			else
 			{
 				// user wants forward control to be forward control
-				if (fabs(fAxisValue) > joy_forwardthreshold.value)
+				if (fabsf(fAxisValue) > joy_forwardthreshold.value)
 				{
 					cmd->forwardmove += (fAxisValue * joy_forwardsensitivity.value) * speed * cl_forwardspeed.value;
 				}
@@ -251,7 +251,7 @@ void IN_JoyMove (usercmd_t *cmd)
 			break;
 
 		case AxisSide:
-			if (fabs(fAxisValue) > joy_sidethreshold.value)
+			if (fabsf(fAxisValue) > joy_sidethreshold.value)
 			{
 				cmd->sidemove += (fAxisValue * joy_sidesensitivity.value) * speed * cl_sidespeed.value;
 			}
@@ -261,7 +261,7 @@ void IN_JoyMove (usercmd_t *cmd)
 			if ((in_strafe.state & 1) || (lookstrafe.value && (in_mlook.state & 1)))
 			{
 				// user wants turn control to become side control
-				if (fabs(fAxisValue) > joy_sidethreshold.value)
+				if (fabsf(fAxisValue) > joy_sidethreshold.value)
 				{
 					cmd->sidemove -= (fAxisValue * joy_sidesensitivity.value) * speed * cl_sidespeed.value;
 				}
@@ -269,7 +269,7 @@ void IN_JoyMove (usercmd_t *cmd)
 			else
 			{
 				// user wants turn control to be turn control
-				if (fabs(fAxisValue) > joy_yawthreshold.value)
+				if (fabsf(fAxisValue) > joy_yawthreshold.value)
 				{
 					if(dwControlMap[i] == JOY_ABSOLUTE_AXIS)
 					{
@@ -287,7 +287,7 @@ void IN_JoyMove (usercmd_t *cmd)
 		case AxisLook:
 			if (in_mlook.state & 1)
 			{
-				if (fabs(fAxisValue) > joy_pitchthreshold.value)
+				if (fabsf(fAxisValue) > joy_pitchthreshold.value)
 				{
 					// pitch movement detected and pitch movement desired by user
 					if(dwControlMap[i] == JOY_ABSOLUTE_AXIS)
