+++ src/sv_phys.c
@@ -748,8 +748,8 @@ int SV_TryUnstick (edict_t *ent, vec3_t oldvel)
 		ent->v. velocity[2] = 0;
 		clip = SV_FlyMove (ent, 0.1, &steptrace);
 
-		if ( fabs(oldorg[1] - ent->v.origin[1]) > 4
-		|| fabs(oldorg[0] - ent->v.origin[0]) > 4 )
+		if ( fabsf(oldorg[1] - ent->v.origin[1]) > 4
+		|| fabsf(oldorg[0] - ent->v.origin[0]) > 4 )
 		{
 //Con_DPrintf ("unstuck!\n");
 			return clip;
@@ -832,8 +832,8 @@ void SV_WalkMove (edict_t *ent)
 // in the clipping hulls
 	if (clip)
 	{
-		if ( fabs(oldorg[1] - ent->v.origin[1]) < 0.03125
-		&& fabs(oldorg[0] - ent->v.origin[0]) < 0.03125 )
+		if ( fabsf(oldorg[1] - ent->v.origin[1]) < 0.03125
+		&& fabsf(oldorg[0] - ent->v.origin[0]) < 0.03125 )
 		{	// stepping up didn't make any progress
 			clip = SV_TryUnstick (ent, oldvel);
 		}
